refactor(perms): move org-scope registry into @octopus/db
Single source of truth so octopus-admin can share the exact registry instead
of maintaining a hand-synced mirror. The module is pure (no prisma import),
exposed via a new '@octopus/db/org-permissions' subpath export so it stays out
of the prisma-instantiating index and safe for client components.

apps/web/lib/org-permissions.ts is now a thin re-export — all existing
@/lib/org-permissions imports are unchanged.
