+++ packages/db/package.json
@@ -4,6 +4,11 @@
   "private": true,
   "main": "src/index.ts",
   "types": "src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./org-permissions": "./src/org-permissions.ts",
+    "./package.json": "./package.json"
+  },
   "scripts": {
     "db:generate": "prisma generate",
     "db:migrate": "prisma migrate dev",
